# refactor: use `BlocksSerializer` to replace `StringBlock` to simplify the serialization