… Cargo.lock
… src/query/service/Cargo.toml
… src/query/service/src/servers/http/v1/http_query_handlers.rs
… src/query/service/src/servers/http/v1/mod.rs
… src/query/service/src/servers/http/v1/query/http_query.rs
… src/query/service/src/servers/http/v1/query/mod.rs
… src/query/service/src/servers/http/v1/query/page_manager.rs
… src/query/service/src/servers/http/v1/query/string_block.rs
… src/query/service/tests/it/servers/http/http_query_handlers.rs
… src/query/service/tests/it/servers/http/json_block.rs