I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

- fixes: https://github.com/databendlabs/databend/issues/17662

when using `StringBlocks` to query large amounts of data, data was frequently materialized and there were meaningless data conversions. For example, the String type was first converted to Bytes and then to String.

This PR uses `BlocksSerializer` to collect `DataBlocks` that need to be serialized. When serde serialization is performed, it will imitate the behavior of Vec for serialization, thus avoiding the need for `DataBlock` to be converted to string line by line in the original `StringBlock` and then collected as `Vec<Vec<Option<String>>>`

When extracting data from a `DataBlock`, `BlocksSerializer` will try to directly extract data from String, Variant, and other types that are Strings, avoiding the meaningless conversion to Bytes in the middle.

Originally, when processing `remaining_size` and `remaining_rows`, the specific size of the data is calculated row by row. However, when the amount of data is large, this will take up some overhead. Therefore, this PR uses the `Proportional Splitting`. Assuming that the data in DataBlock is uniform, `take_rows = (remain_size * block_rows) / block_size` is used to determine how many rows of data can be obtained under the `remaining_size` limit, and then further limit it through `remaining_row`.

test case: 
```sql
CREATE TABLE test AS
SELECT 
    number AS id,
    JSON_OBJECT(
        'id', CAST(number % 100000000 AS STRING),
        'event_type', CONCAT('submit_abcd', CAST(number % 100 AS STRING)),
        'client_code', CONCAT('aaaaaaaaaaaaaaaaaaa', CAST(number % 100 AS STRING)),
        'activity_id', CONCAT('aaaaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbbbbbbbbccccccccccccccddddddddddddddddeeeeeeeeeeeeeeeee', CAST(number % 100 AS STRING))
    ) AS payload
FROM numbers(50000000);


SELECT * FROM test;
```

Release Mode: 
- Before: 50000000 rows read in 66.891 sec. Processed 50 million rows, 10.19 GiB (747.48 thousand rows/s, 155.92 MiB/s)
- After: 50000000 rows read in 53.671 sec. Processed 50 million rows, 10.19 GiB (931.6 thousand rows/s, 194.33 MiB/s)

![e132dab7e5a558b90f124541436ed401](https://github.com/user-attachments/assets/18f14616-ec00-40e3-9f0a-e110fb7b94cd)

[after.pb.gz](https://github.com/user-attachments/files/19501389/after.pb.gz)
[before.pb.gz](https://github.com/user-attachments/files/19501390/before.pb.gz)

## Tests

- [ ] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [x] No Test - There are already tests covering this: `test_max_size_per_page` & `test_max_size_per_page_total_rows`

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/databendlabs/databend/17667)
<!-- Reviewable:end -->